Filesystem iosocket #10

+++ server/server.py
@@ -1,5 +1,6 @@
-from flask import Flask, render_template, request, jsonify
-from flask_socketio import SocketIO, send, join_room, leave_room
+from flask import Flask, render_template, request, jsonify, session
+from flask_session import Session
+from flask_socketio import SocketIO, send, join_room, leave_room, close_room
 from model import db, admin, User, Planning_group
 import config
 from helpers import *
@@ -21,13 +22,17 @@ app = Flask(__name__, static_folder=static_folder_path, template_folder=template
 app.config['SQLALCHEMY_DATABASE_URI'] = db_uri
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False # To supress a warning
 app.config['SECRET_KEY'] = config.MY_SECRET
+app.config['SESSION_TYPE'] = 'filesystem'
+app.config['SESSION_COOKIE_NAME'] = 'session_id'
 
-sio = SocketIO()
+Session(app)
+sio = SocketIO(manage_session=False)
+
 
 """ dev only """
 from flask_cors import CORS, cross_origin
-cors = CORS(app)
 app.config['CORS_HEADERS'] = 'Content-Type'
+cors = CORS(app, supports_credentials=True)
 """ end """
 
 def create_prod_app(app):
@@ -50,13 +55,26 @@ def create_test_app(app):
     return app
 
 """ SOCKET IO """
-@sio.on('message')
-def handleMessage(msg):
-    print('Message: ' + msg)
-    send(msg, broadcast=True)
+# @sio.on('message')
+# def handleMessage(msg):
+#     print('Message: ' + msg)
+#     send(msg, broadcast=True)
 
+@sio.on('connect')
+def connect_user():
+    session['sid'] = request.sid
+    session['namespace'] = request.namespace
+    print(session)
 
 """ API ENDPOINTS """
+
+@app.route('/api/test', methods=['POST'])
+def test():
+    print("IN HERE")
+    print(session)
+    session['test'] = request.json['test']
+    return jsonify("test"), 200
+
 # Example payload
 # {
 # 	"group_name":"test_group",
@@ -70,7 +88,6 @@ def handleMessage(msg):
 #   "id_token":"googles long token id in response.getAuthResponse().id_token"
 # }
 @app.route('/api/creategroup', methods=['POST'])
-@cross_origin() #dev only
 def create_group():
     """ Creates a group with the owner being the user
     that is included in the POST. 
@@ -81,6 +98,8 @@ def create_group():
     """
     try:
         with handle_exceptions():
+            sid = session['sid']
+            namespace = session['namespace']
             payload = request.json
             if payload is None:
                 raise ValueError("Missing json body in post")
@@ -119,6 +138,8 @@ def create_group():
             db.session.commit()
             new_group.users.append(user)
             db.session.commit()
+            # Join the socket-io room 
+            join_room(new_group.group_str_id, sid=sid, namespace=namespace)
             return jsonify({
                 'group_str_id': new_group.group_str_id,
                 'google_id': user.google_id
@@ -138,7 +159,6 @@ def create_group():
 #   "id_token":"googles long token id in response.getAuthResponse().id_token"
 # }
 @app.route('/api/adduser', methods=['POST'])
-@cross_origin() #dev only
 @require_group_str_id
 def add_user(group=None):
     """ Adds a user to the group identified by
@@ -149,13 +169,23 @@ def add_user(group=None):
     """
     try:
         with handle_exceptions():
+            if 'sid' not in session or 'namespace' not in session:
+                raise ValueError("Missing session values")
+            sid = session['sid']
+            namespace = session['namespace']
             payload = request.json
+            if payload is None:
+                raise ValueError("Missing json body in post")
             name = payload['name']
             id_token = payload['id_token']
             access_token = payload['access_token']
             user = create_or_find_user(id_token, name, access_token)
             group.users.append(user)
             db.session.commit()
+            # Join the socket-io room group_str_id
+            join_room(group.group_str_id, sid=sid, namespace=namespace)
+            # Notify all in that room
+            send("New user joined", room=group.group_str_id, sid=sid, namespace=namespace)
             return jsonify({'google_id': user.google_id}), 200
     except APIError as e:
         return e.response, e.code
@@ -167,7 +197,6 @@ def add_user(group=None):
 #   "google_id":"user google id"
 # }
 @app.route('/api/getusersfromgroup')
-@cross_origin() # dev only
 @require_login
 @require_group_str_id
 def get_users_from_group(user=None, group=None):
@@ -195,7 +224,6 @@ def get_users_from_group(user=None, group=None):
 #   "google_id":"user google id"
 # }
 @app.route('/api/getgroupcalendar')
-@cross_origin() # dev only
 @require_login
 @require_group_str_id
 def get_group_calendar(user=None, group=None):
@@ -218,7 +246,6 @@ def get_group_calendar(user=None, group=None):
 #   "google_id":"user google id"
 # }
 @app.route('/api/remove', methods=['DELETE'])
-@cross_origin() # dev only
 @require_login
 @require_group_str_id
 def remove(user=None, group=None):
@@ -231,14 +258,22 @@ def remove(user=None, group=None):
     """
     try:
         with handle_exceptions():
+            if 'sid' not in session or 'namespace' not in session:
+                raise ValueError("Missing session values")
+            sid = session['sid']
+            namespace = session['namespace']
             if group.owner.google_id == user.google_id:
                 for g_user in group.users:
                     group.users.remove(g_user)
                     attempt_delete_user(g_user)
                 db.session.delete(group)
+                # Close socket-io room 
+                close_room(group.group_str_id, sid=sid, namespace=namespace)
             else:
                 group.users.remove(user)
                 attempt_delete_user(user)
+                # Leave socket-io room
+                leave_room(group.group_str_id, sid=sid, namespace=namespace)
             db.session.commit()
             return "", 202
     except APIError as e:
@@ -270,7 +305,6 @@ def update_access_token(user=None):
 # Has to be last route!
 @app.route('/', defaults={'path': ''})
 @app.route('/<path:path>')
-@cross_origin() #dev only
 def index(path):
     """ Renders the actual react webpage.
     """
@@ -282,4 +316,5 @@ if __name__ == '__main__':
     # with app.app_context():
     #     db.drop_all()
     #     db.create_all()
-    app.run(port=5000, debug=True)
+    # app.run(port=5000, debug=True)
+    sio.run(app, port=5000, debug=True)
